Untangle some transitive imports
