+++ lib/tether/tether.go
@@ -37,6 +37,7 @@ import (
 
 	"github.com/vmware/vic/cmd/tether/msgs"
 	"github.com/vmware/vic/lib/config/executor"
+	"github.com/vmware/vic/lib/tether/shared"
 	"github.com/vmware/vic/lib/system"
 	"github.com/vmware/vic/pkg/dio"
 	"github.com/vmware/vic/pkg/log/syslog"
@@ -59,7 +60,7 @@ const (
 // Sys is used to configure where the target system files are
 var (
 	// Used to access the acutal system paths and files
-	Sys = system.New()
+	Sys = shared.Sys
 	// Used to access and manipulate the tether modified bind sources
 	// that are mounted over the system ones.
 	BindSys = system.NewWithRoot("/.tether")
@@ -164,14 +165,16 @@ func (t *tether) setup() error {
 		}
 	}
 
+	pidDir := shared.PIDFileDir()
+
 	// #nosec: Expect directory permissions to be 0700 or less
-	if err = os.MkdirAll(PIDFileDir(), 0755); err != nil {
-		log.Errorf("could not create pid file directory %s: %s", PIDFileDir(), err)
+	if err = os.MkdirAll(pidDir, 0755); err != nil {
+		log.Errorf("could not create pid file directory %s: %s", pidDir, err)
 	}
 
 	// Create PID file for tether
 	tname := path.Base(os.Args[0])
-	err = ioutil.WriteFile(fmt.Sprintf("%s.pid", path.Join(PIDFileDir(), tname)),
+	err = ioutil.WriteFile(fmt.Sprintf("%s.pid", path.Join(pidDir, tname)),
 		[]byte(fmt.Sprintf("%d", os.Getpid())),
 		0644)
 	if err != nil {
@@ -655,8 +658,8 @@ func (t *tether) handleSessionExit(session *SessionConfig) {
 
 	// Remove associated PID file
 	cmdname := path.Base(session.Cmd.Path)
-	// #nosec: Errors unhandled.
-	_ = os.Remove(fmt.Sprintf("%s.pid", path.Join(PIDFileDir(), cmdname)))
+
+	_ = os.Remove(fmt.Sprintf("%s.pid", path.Join(shared.PIDFileDir(), cmdname)))
 
 	// set the stop time
 	session.StopTime = time.Now().UTC().Unix()
@@ -817,7 +820,7 @@ func (t *tether) launch(session *SessionConfig) error {
 
 	// Write the PID to the associated PID file
 	cmdname := path.Base(session.Cmd.Path)
-	err = ioutil.WriteFile(fmt.Sprintf("%s.pid", path.Join(PIDFileDir(), cmdname)),
+	err = ioutil.WriteFile(fmt.Sprintf("%s.pid", path.Join(shared.PIDFileDir(), cmdname)),
 		[]byte(fmt.Sprintf("%d", pid)),
 		0644)
 	if err != nil {
@@ -934,10 +937,6 @@ func (t *tether) Flush() error {
 
 	extraconfig.Encode(t.sink, t.config)
 	return nil
-}
-
-func PIDFileDir() string {
-	return path.Join(Sys.Root, pidFilePath)
 }
 
 // killHelper was pulled from toolbox, and that variant should be directed at this
